Get data from a local csv file

Adjusted version of read.csv() to conveniently load WATLAS data from a csv file. To read large csv files efficiently, the functions uses the package data.table but the output is converted to a data frame.

Usage

atl_get_data_csv(
  fpath = system.file("extdata", "redknot_2707_WATLAS_exampledata.csv", package =
    "tools4watlas")
)

Arguments

fpath

The full path (including path and file name) of the csv file. If left empty, the example data file supplied with tools4watlas will be used.

Value

A formatted dataframe of the csv file with the columns:
PosID = Unique number for localizations
TAG = 11 digit WATLAS tag ID (numeric)
tag = 4 digit tag number (character)
TIME = UNIX time (seconds)
time = Timestamp in POSIXct (UTC)
X = X-ccordinates in meters (utm 31 N)
Y = Y-ccordinates in meters (utm 31 N)
NBS = Number of Base Stations used in calculating coordinates
VARX = Variance in estimating X-coordinates
VARY = Variance in estimating Y-coordinates
COVXY = Co-variance between X- and Y-coordinates

Author

Allert Bijleveld