WATLAS species labels

Returns a named vector of species labels in either a multiline or single-line format.

Usage

atl_spec_labs(option = "multiline")

Arguments

option

A character string specifying the format of the species names. Options are "multiline" (default), where names include line breaks (\n), or "singleline", where names are returned as a single line.

Value

A named character vector where names correspond to species identifiers and values are formatted species names.

Examples

library(tools4watlas)
atl_spec_labs("multiline")
#>                    curlew         bar-tailed godwit             oystercatcher 
#>        "Eurasian\ncurlew"      "Bar-tailed\ngodwit" "Eurasian\noystercatcher" 
#>                  redshank                  red knot                sanderling 
#>        "Common\nredshank"                "Red knot"              "Sanderling" 
#>                    dunlin                 turnstone               grey plover 
#>                  "Dunlin"               "Turnstone"            "Grey\nplover" 
#>          curlew sandpiper                 spoonbill            kentish plover 
#>       "Curlew\nsandpiper"     "Eurasian\nspoonbill"         "Kentish\nplover" 
atl_spec_labs("multiline")
#>                    curlew         bar-tailed godwit             oystercatcher 
#>        "Eurasian\ncurlew"      "Bar-tailed\ngodwit" "Eurasian\noystercatcher" 
#>                  redshank                  red knot                sanderling 
#>        "Common\nredshank"                "Red knot"              "Sanderling" 
#>                    dunlin                 turnstone               grey plover 
#>                  "Dunlin"               "Turnstone"            "Grey\nplover" 
#>          curlew sandpiper                 spoonbill            kentish plover 
#>       "Curlew\nsandpiper"     "Eurasian\nspoonbill"         "Kentish\nplover"