Check the residency patches from one tag during one tide

Generates a ggplot2 showing bird residency patches per tideID, including movement paths, patch durations, and an inset overview map.

Usage

atl_check_res_patch(
  tag,
  tide,
  data,
  patch_data,
  tide_data,
  buffer = 250,
  buffer_overview = 10000,
  point_size = 1,
  point_alpha = 0.5,
  path_linewidth = 0.5,
  path_alpha = 0.2,
  patch_label_size = 4,
  patch_label_padding = 1,
  element_text_size = 11,
  filename = NULL,
  png_width = 3840,
  png_height = 2160
)

Arguments

tag

Bird tag ID to subset.

tide

Tide ID to subset.

data

A data.table containing tracking data. Must include the columns: "tag", "x", "y", "time","datetime", and species.

patch_data

A data.table containing residency patch details as created by atl_res_patch().

tide_data

Data on the timing (in UTC) of low and high tides.

buffer

Map buffer size (default: 250).

buffer_overview

Overview map buffer size (default: 10000).

point_size

Size of plotted points (default: 1).

point_alpha

Transparency of points (default: 0.5).

path_linewidth

Line width of movement paths (default: 0.5).

path_alpha

Transparency of movement paths (default: 0.2).

patch_label_size

Font size for patch labels (default: 4).

patch_label_padding

Padding for patch labels (default: 1).

element_text_size

Font size for axis and legend text (default: 11).

filename

Character (or NULL). If provided, the plot is saved as a .png file to this path and with this name; otherwise, the function returns the plot.

png_width

Width of saved PNG (default: 3840).

png_height

Height of saved PNG (default: 2160).

Value

A ggplot object or a saved PNG file.

Author

Johannes Krietsch